Fix renaming FunctionReturnDecoder and FunctionReturnDecoderProvider
